Add "Projects" menu item and navigation functionality
Added a new "Projects" menu item to the off-canvas menu in
`menu-offcanvas.component.html`, styled similarly to the
existing "Chats" menu item. Clicking this item navigates
to the "Projects" page.

Implemented the `onClickGoToProjects()` method in
`menu-offcanvas.component.ts` to handle navigation to the
"projects" route using the Angular Router.

Updated method documentation to describe the purpose and
behavior of `onClickGoToProjects()`.